==> Train_model/03-Model_optimisation_hyperparameters_turning.ipynb <==
import pandas as pd
import numpy as np

from sklearn.model_selection import train_test_split
import xgboost as xgb
from sklearn.metrics import accuracy_score, classification_report, confusion_matrix

from sklearn.metrics import recall_score, make_scorer

from sklearn.model_selection import GridSearchCV
from sklearn.model_selection import RandomizedSearchCV

import json


### Download XGBoost model with selected features for optimisation

data = pd.read_csv('../Project_datasets/reduced_dataset.csv')

X = data.drop(['purchase'], axis=1)
y = data['purchase']

# Split the data into training and testing sets
# This splits into 70% train, 30% test by default
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

# Create XGBoost classifier
model = xgb.XGBClassifier(random_state=42)

# Train model
model.fit(X_train, y_train)

# Make predictions
y_pred = model.predict(X_test)

# Evaluate
accuracy = accuracy_score(y_test, y_pred)
print(f"\nAccuracy with XGBoost importance selected features: {accuracy:.4f}")

print("\nClassification Report:")
print(classification_report(y_test, y_pred))

print("\nConfusion Matrix:")
print(confusion_matrix(y_test, y_pred))


### Hyperparamenters turning

# Custom scorer that focuses only on recall for class 1 (responders)
custom_recall_scorer = make_scorer(recall_score, pos_label=1)

# Parameter grid -- random generated numbers
param = {
    'max_depth': [3, 6, 8],
    'learning_rate': [0.05, 0.1, 0.2],
    'n_estimators': [100, 200, 300],
    'subsample': [0.8, 1.0],
    'colsample_bytree': [0.8, 1.0],
    'gamma': [0, 1],
    'min_child_weight': [1, 3, 5]
}

###### GridSearchCV

# GridSearchCV with the custom recall scorer
grid_search = GridSearchCV(
    estimator=model,
    param_grid=param,
    scoring=custom_recall_scorer,
    cv=3,
    verbose=1,
    n_jobs=-1
)

# Fit to training data
grid_search.fit(X_train, y_train)

# Get the best model
best_model = grid_search.best_estimator_

# Accuracy
y_pred = best_model.predict(X_test)
print("Best Parameters:", grid_search.best_params_)
print("\nClassification Report:")
print(classification_report(y_test, y_pred))
print("\nConfusion Matrix:")
print(confusion_matrix(y_test, y_pred))

###### RandomizedSearchCV

# RandomizedSearchCV ith the custom recall scorer

random_search = RandomizedSearchCV(
    estimator=model,
    param_distributions=param,
    n_iter=30,  
    scoring=custom_recall_scorer,
    cv=3,
    verbose=1,
    random_state=42,
    n_jobs=-1
)

# Fit the randomized search
random_search.fit(X_train, y_train)

#  Best model
best_model = random_search.best_estimator_

# Accuracy

y_pred = best_model.predict(X_test)
print("Best Parameters:", random_search.best_params_)
print("\nClassification Report:")
print(classification_report(y_test, y_pred))
print("\nConfusion Matrix:")
print(confusion_matrix(y_test, y_pred))


#### Chose the Search and save into dict

with open('../Project_datasets/best_xgb_params.json', 'w') as f:
    json.dump(grid_search.best_params_, f)
