==> Train_model/04-Model_optimisation_cross_validation.ipynb <==
import pandas as pd
import numpy as np
import json

from sklearn.model_selection import train_test_split
import xgboost as xgb
from sklearn.metrics import accuracy_score, classification_report, confusion_matrix
from sklearn.metrics import recall_score

from sklearn.model_selection import StratifiedKFold
from sklearn.model_selection import ShuffleSplit




### Download data and hyperparameters, run initial model

data = pd.read_csv('../Project_datasets/reduced_dataset.csv')

X = data.drop(['purchase'], axis=1)
y = data['purchase']


with open('../Project_datasets/best_xgb_params.json', 'r') as f:
    best_params = json.load(f)


# Split the data into training and testing sets
# This splits into 70% train, 30% test by default
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

model = xgb.XGBClassifier(**best_params)
model.fit(X_train, y_train)

# Make predictions
y_pred = model.predict(X_test)

# Evaluate
accuracy = accuracy_score(y_test, y_pred)
print(f"\nAccuracy with XGBoost importance selected features: {accuracy:.4f}")

print("\nClassification Report:")
print(classification_report(y_test, y_pred))

print("\nConfusion Matrix:")
print(confusion_matrix(y_test, y_pred))


### Cross-validation

###### 10 folders with different train-test split ration

# Create StratifiedKFold model
skf = StratifiedKFold(n_splits=10, shuffle=True, random_state=42)

accuracies = []

# 10-fold cross-validation
for fold, (train_index, test_index) in enumerate(skf.split(X, y), 1):
    
    X_train, X_test = X.iloc[train_index], X.iloc[test_index]
    y_train, y_test = y.iloc[train_index], y.iloc[test_index]

    # Model train
    model.fit(X_train, y_train)
    
    
    model.fit(X_train, y_train)

    # Predictions
    y_train_pred = model.predict(X_train)
    y_test_pred = model.predict(X_test)

    # Evaluation
    print(f"\n--- Fold {fold} ---")
    print("Train recall:", recall_score(y_train, y_train_pred, pos_label=1))
    print("Test recall :", recall_score(y_test, y_test_pred, pos_label=1))




###### 10 folders with different data samples with constant train-test split 70/30

# Random 70/30 split
rs = ShuffleSplit(n_splits=10, test_size=0.3, train_size=0.7, random_state=42)

train_recalls = []
test_recalls = []

for i, (train_idx, test_idx) in enumerate(rs.split(X, y), 1):
    X_train, X_test = X.iloc[train_idx], X.iloc[test_idx]
    y_train, y_test = y.iloc[train_idx], y.iloc[test_idx]
    
    model.fit(X_train, y_train)
    
    y_train_pred = model.predict(X_train)
    y_test_pred = model.predict(X_test)
    
    train_recall = recall_score(y_train, y_train_pred, pos_label=1)
    test_recall = recall_score(y_test, y_test_pred, pos_label=1)
    
    train_recalls.append(train_recall)
    test_recalls.append(test_recall)
    
    print(f"\n--- Bucket {i} ---")
    print(f"Train recall: {train_recall:.4f}")
    print(f"Test recall : {test_recall:.4f}")

# Summary
print(f"\nAverage Train Recall: {np.mean(train_recalls):.4f}")
print(f"Average Test Recall : {np.mean(test_recalls):.4f}")


# Save the model to file
model.save_model("../Models/xgb_model.json")
