==> Train_model/05-Price_optization.ipynb <==
import pandas as pd
import numpy as np
import json
import xgboost as xgb

# simulate data  

data = pd.read_csv('../Project_datasets/reduced_dataset.csv')
users_list = data.drop(['purchase'], axis = 1)

# Load the model from file
purchase_model = xgb.XGBClassifier()
purchase_model.load_model("../Models/xgb_model.json")


# Define parameters
discount_range = np.arange(0, 0.30, 0.05)  # 0% to 50% in 1% steps
probability_threshold = 0.6

#  Model time optimisation for the big users batch

#  Containers
assigned_discount = pd.Series(np.nan, index=users_df.index, name="optimal_discount")
assigned_probability = pd.Series(np.nan, index=users_df.index, name="prob_at_optimal")

#  Feature + discount grid 
def prepare_batch(features: pd.DataFrame, discount: float) -> pd.DataFrame:
    temp = features.copy()
    temp["discount%"] = discount
    temp = pd.get_dummies(temp)
    return temp


remaining_idx = users_df.index

#  Run the users batch through the incremented discount loop 
for d in discount_levels:
    if remaining_idx.size == 0:
        break

    user_batch = prepare_batch(users_df.loc[remaining_idx], d)
    probs = purchase_model.predict_proba(user_batch)[:, 1]

    meets = probs >= probability_threshold
    if not meets.any():
        continue

    qualifying_idx = remaining_idx[meets]
    assigned_discount.loc[qualifying_idx] = d
    assigned_probability.loc[qualifying_idx] = probs[meets]

    # Keep only users that haven't yet qualified
    remaining_idx = remaining_idx[~meets]

# Assemble result
result = users_list.copy()
result["optimal_discount"] = assigned_discount
result["prob_at_optimal"] = assigned_probability

# Qualified users only
final_result = result.loc[result["optimal_discount"].notna(),
                          ["optimal_discount", "prob_at_optimal"]]



final_result
